rename readers in code
